pyproject.toml
[project]
name = "compilador"
version = "0.1.0"
description = "Compilador escrito em python"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
    "cyclopts>=3.13.1",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "pytest>=8.3.5",
    "pytest-sugar>=1.0.0",
    "ruff>=0.11.6",
]

[project.scripts]
compilador = "compilador.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
exclude = ["compilador/parser/*"]

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
